antigravity_agent/requirements.txt
langgraph
langchain
langchain-community
langchain-google-genai
chromadb
pydantic
tiktoken
python-dotenv
DDGS
tavily-python